remove unnecessary casts, checks etc

+++ src/main/java/mekanism/client/sound/TileSound.java
@@ -70,12 +70,12 @@ public class TileSound extends MekSound implements IResettableSound
 
 	public float getFadeInMultiplier() 
 	{
-		return ticks >= fadeIn ? 1 : (float)(ticks / (float)fadeIn);
+		return ticks >= fadeIn ? 1 : (ticks / (float)fadeIn);
 	}
 
 	public float getFadeOutMultiplier() 
 	{
-		return ticks >= fadeOut ? 0 : (float)((fadeOut - ticks) / (float)fadeOut);
+		return ticks >= fadeOut ? 0 : ((fadeOut - ticks) / (float)fadeOut);
 	}
 
 	/* ITickableSound */
